# specs/peixian-data-adapter-openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "沛县公安统一数据适配服务",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/person/profile/query": {
      "post": {
        "summary": "Person Profile",
        "operationId": "person_profile_v1_person_profile_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_PersonProfile_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/person/houses/query": {
      "post": {
        "summary": "Person Houses",
        "operationId": "person_houses_v1_person_houses_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_HouseRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/person/companies/query": {
      "post": {
        "summary": "Person Companies",
        "operationId": "person_companies_v1_person_companies_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_CompanyRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/person/family/query": {
      "post": {
        "summary": "Person Family",
        "operationId": "person_family_v1_person_family_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_FamilyRelation_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/police/cases/query": {
      "post": {
        "summary": "Police Cases",
        "operationId": "police_cases_v1_police_cases_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_CaseRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/police/incidents/query": {
      "post": {
        "summary": "Police Incidents",
        "operationId": "police_incidents_v1_police_incidents_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_IncidentRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/police/disputes/query": {
      "post": {
        "summary": "Police Disputes",
        "operationId": "police_disputes_v1_police_disputes_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_DisputeRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mobility/tracks/query": {
      "post": {
        "summary": "Mobility Tracks",
        "operationId": "mobility_tracks_v1_mobility_tracks_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_TrackRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mobility/hotels/query": {
      "post": {
        "summary": "Mobility Hotels",
        "operationId": "mobility_hotels_v1_mobility_hotels_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_HotelRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mobility/railway/query": {
      "post": {
        "summary": "Mobility Railway",
        "operationId": "mobility_railway_v1_mobility_railway_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_RailwayRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mobility/netbar/query": {
      "post": {
        "summary": "Mobility Netbar",
        "operationId": "mobility_netbar_v1_mobility_netbar_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_NetbarRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/vehicle/motor/query": {
      "post": {
        "summary": "Vehicle Motor",
        "operationId": "vehicle_motor_v1_vehicle_motor_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_MotorVehicleRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/vehicle/non-motor/query": {
      "post": {
        "summary": "Vehicle Non Motor",
        "operationId": "vehicle_non_motor_v1_vehicle_non_motor_query_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse_NonMotorVehicleRecord_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageRequest": {
        "properties": {
          "number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Number",
            "default": 1
          },
          "size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Size",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PageRequest"
      },
...
      "PersonQuery": {
        "properties": {
          "certificate_no": {
            "type": "string",
            "pattern": "^[0-9]{17}[0-9Xx]$",
            "title": "Certificate No"
          },
          "begin_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Begin Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "page": {
            "$ref": "#/components/schemas/PageRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "certificate_no"
        ],
        "title": "PersonQuery"
      },
      "QueryResponse_CaseRecord_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "returned_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Returned Count"
          },
          "total_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Count"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "queried_at",
          "returned_count",
          "total_count",
          "page",
          "items"
        ],
        "title": "QueryResponse[CaseRecord]"
      },
      "QueryResponse_CompanyRecord_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "returned_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Returned Count"
          },
          "total_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Count"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanyRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "queried_at",
          "returned_count",
          "total_count",
          "page",
          "items"
        ],
        "title": "QueryResponse[CompanyRecord]"
      },
...
      "QueryResponse_FamilyRelation_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "returned_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Returned Count"
          },
          "total_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Count"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FamilyRelation"
            },
            "type": "array",
            "title": "Items"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "queried_at",
          "returned_count",
          "total_count",
          "page",
          "items"
        ],
        "title": "QueryResponse[FamilyRelation]"
      },
...
      "QueryResponse_HouseRecord_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "returned_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Returned Count"
          },
          "total_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Count"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HouseRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "queried_at",
          "returned_count",
          "total_count",
          "page",
          "items"
        ],
        "title": "QueryResponse[HouseRecord]"
      },
...
      "QueryResponse_PersonProfile_": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "queried_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queried At"
          },
          "returned_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Returned Count"
          },
          "total_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Count"
          },
          "page": {
            "$ref": "#/components/schemas/PageResult"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PersonProfile"
            },
            "type": "array",
            "title": "Items"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trace_id",
          "source",
          "queried_at",
          "returned_count",
          "total_count",
          "page",
          "items"
        ],
        "title": "QueryResponse[PersonProfile]"
      },
...
    }
  }
}